services/workflow/pyproject.toml
[project]
name = "workflow"
version = "0.1.0"
description = "Airflow workflow orchestration service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
# Enable pycodestyle (E), Pyflakes (F), and isort (I)
select = ["E", "F", "I", "UP", "N", "S", "B", "A", "C4", "DTZ", "PIE", "SIM", "PTH"]
ignore = [
    "S101",   # Allow assert statements (needed for pytest)
    "DTZ005", # Allow datetime.now() without timezone (common in Airflow)
    "DTZ001", # Allow datetime without tzinfo in start_date (Airflow handles this)
    "S603",   # Allow subprocess calls (common in Airflow tasks)
    "S607",   # Allow partial executable paths
    "T201",   # Allow print statements (common in Airflow tasks for logging)
    "B018",   # Allow useless expressions (DAG task definition pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["dags", "tests"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, can tighten later
ignore_missing_imports = true  # Airflow has incomplete type stubs
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
